+++ src/containers/App/saga.js
@@ -1,5 +1,12 @@
-import { all, call, put, take, takeEvery, takeLatest } from 'redux-saga/effects';
-import { push } from 'react-router-redux';
+import {
+  all,
+  call,
+  put,
+  take,
+  takeEvery,
+  takeLatest,
+} from 'redux-saga/effects';
+import { push } from 'connected-react-router';
 import request from 'utils/request';
 
 import { authCall } from 'shared/services/api/api';
@@ -11,7 +18,7 @@ import {
   LOGIN,
   AUTHENTICATE_USER,
   REQUEST_CATEGORIES,
-  UPLOAD_REQUEST
+  UPLOAD_REQUEST,
 } from './constants';
 import {
   showGlobalError,
@@ -19,7 +26,7 @@ import {
   requestCategoriesSuccess,
   uploadProgress,
   uploadSuccess,
-  uploadFailure
+  uploadFailure,
 } from './actions';
 import { login, logout, getOauthDomain } from '../../shared/services/auth/auth';
 
@@ -40,7 +47,12 @@ export function* callLogout() {
   try {
     // This forces the remove of the grip cookies.
     if (getOauthDomain() === 'grip') {
-      window.open('https://auth.grip-on-it.com/v2/logout?tenantId=rjsfm52t', '_blank').close();
+      window
+        .open(
+          'https://auth.grip-on-it.com/v2/logout?tenantId=rjsfm52t',
+          '_blank',
+        )
+        .close();
     }
     logout();
     yield put(push('/'));
@@ -56,7 +68,11 @@ export function* callAuthorize(action) {
       const requestURL = `${baseUrl}`;
 
       const user = yield authCall(requestURL, null, accessToken);
-      const credentials = { ...action.payload, userScopes: [...user.groups], userPermissions: [...user.permissions] };
+      const credentials = {
+        ...action.payload,
+        userScopes: [...user.groups],
+        userPermissions: [...user.permissions],
+      };
       yield put(authorizeUser(credentials));
     }
   } catch (error) {
@@ -65,7 +81,9 @@ export function* callAuthorize(action) {
 }
 
 export function* fetchCategories() {
-  const requestURL = `${CONFIGURATION.API_ROOT}signals/v1/public/terms/categories/`;
+  const requestURL = `${
+    CONFIGURATION.API_ROOT
+  }signals/v1/public/terms/categories/`;
 
   try {
     const categories = yield call(request, requestURL);
@@ -83,7 +101,12 @@ export function* uploadFileWrapper(action) {
 export function* uploadFile(action) {
   const requestURL = `${CONFIGURATION.API_ROOT}signals/signal/image/`;
 
-  const channel = yield call(fileUploadChannel, requestURL, action.payload.file, action.payload.id);
+  const channel = yield call(
+    fileUploadChannel,
+    requestURL,
+    action.payload.file,
+    action.payload.id,
+  );
   const forever = true;
   while (forever) {
     const { progress = 0, error, success } = yield take(channel);
@@ -106,6 +129,6 @@ export default function* watchAppSaga() {
     takeLatest(LOGOUT, callLogout),
     takeLatest(AUTHENTICATE_USER, callAuthorize),
     takeLatest(REQUEST_CATEGORIES, fetchCategories),
-    takeEvery(UPLOAD_REQUEST, uploadFileWrapper)
+    takeEvery(UPLOAD_REQUEST, uploadFileWrapper),
   ]);
 }