+++ src/signals/incident-management/containers/IncidentSplitContainer/saga.js
@@ -1,5 +1,5 @@
 import { all, put, takeLatest } from 'redux-saga/effects';
-import { push } from 'react-router-redux';
+import { push } from 'connected-react-router';
 
 import CONFIGURATION from 'shared/services/configuration/configuration';