+++ src/signals/incident/containers/IncidentContainer/saga.js
@@ -1,5 +1,5 @@
 import { all, call, put, select, takeLatest } from 'redux-saga/effects';
-import { replace } from 'react-router-redux';
+import { replace } from 'connected-react-router';
 import request from 'utils/request';
 
 import { authPostCall } from 'shared/services/api/api';