+++ src/utils/injectSaga.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import hoistNonReactStatics from 'hoist-non-react-statics';
+import { ReactReduxContext } from 'react-redux';
 
 import getInjectors from './sagaInjectors';
 
@@ -9,33 +9,35 @@ import getInjectors from './sagaInjectors';
  *
  * @param {string} key A key of the saga
  * @param {function} saga A root saga that will be injected
- * @param {string} [mode] By default (constants.RESTART_ON_REMOUNT) the saga will be started on component mount and
- * cancelled with `task.cancel()` on component un-mount for improved performance. Another two options:
- *   - constants.DAEMON—starts the saga on component mount and never cancels it or starts again,
- *   - constants.ONCE_TILL_UNMOUNT—behaves like 'RESTART_ON_REMOUNT' but never runs it again.
+ * @param {string} [mode] By default (constants.DAEMON) the saga will be started
+ * on component mount and never canceled or started again. Another two options:
+ *   - constants.RESTART_ON_REMOUNT — the saga will be started on component mount and
+ *   cancelled with `task.cancel()` on component unmount for improved performance,
+ *   - constants.ONCE_TILL_UNMOUNT — behaves like 'RESTART_ON_REMOUNT' but never runs it again.
  *
  */
+/* eslint-disable no-undef */
 export default ({ key, saga, mode }) => (WrappedComponent) => {
   class InjectSaga extends React.Component {
     static WrappedComponent = WrappedComponent;
-    static displayName = `withSaga(${(WrappedComponent.displayName || WrappedComponent.name || 'Component')})`;
-    static contextTypes = {
-      store: PropTypes.object.isRequired,
-    };
 
-    componentWillMount() {
-      const { injectSaga } = this.injectors;
+    static contextType = ReactReduxContext;
 
-      injectSaga(key, { saga, mode }, this.props);
+    static displayName = `withSaga(${WrappedComponent.displayName ||
+      WrappedComponent.name ||
+      'Component'})`;
+
+    constructor(props, context) {
+      super(props, context);
+
+      this.injectors = getInjectors(context.store);
+
+      this.injectors.injectSaga(key, { saga, mode }, this.props);
     }
 
     componentWillUnmount() {
-      const { ejectSaga } = this.injectors;
-
-      ejectSaga(key);
+      this.injectors.ejectSaga(key);
     }
-
-    injectors = getInjectors(this.context.store);
 
     render() {
       return <WrappedComponent {...this.props} />;
@@ -44,3 +46,17 @@ export default ({ key, saga, mode }) => (WrappedComponent) => {
 
   return hoistNonReactStatics(InjectSaga, WrappedComponent);
 };
+
+const useInjectSaga = ({ key, saga, mode }) => {
+  const context = React.useContext(ReactReduxContext);
+  React.useEffect(() => {
+    const injectors = getInjectors(context.store);
+    injectors.injectSaga(key, { saga, mode });
+
+    return () => {
+      injectors.ejectSaga(key);
+    };
+  }, []);
+};
+
+export { useInjectSaga };